Reorganizes Console tests
