--- tests/Feature/Console/Commands/ConfigOrchestrateProviderCommandTest.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace Payavel\Orchestration\Tests\Feature\Console\Commands;
-
-use Payavel\Orchestration\Tests\Traits\CreatesConfigServiceables;
-use Payavel\Orchestration\Tests\Traits\SetsConfigDriver;
-
-class ConfigOrchestrateProviderCommandTest extends TestOrchestrateProviderCommand
-{
-    use CreatesConfigServiceables;
-    use SetsConfigDriver;
-}
