--- tests/Feature/Console/Commands/DatabaseOrchestrateProviderCommandTest.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace Payavel\Orchestration\Tests\Feature\Console\Commands;
-
-use Payavel\Orchestration\Tests\Traits\CreatesDatabaseServiceables;
-use Payavel\Orchestration\Tests\Traits\SetsDatabaseDriver;
-
-class DatabaseOrchestrateProviderCommandTest extends TestOrchestrateProviderCommand
-{
-    use CreatesDatabaseServiceables;
-    use SetsDatabaseDriver;
-}
